+++ backend/src/repositories/dtos/Transaction/TransactionInput.ts
@@ -1,4 +1,4 @@
-import { operators } from '..';
+import { operators } from '../../helpers';
 import { TransactionModelDTO } from './TransactionModel';
 
 export type TransactionInputDTO = Pick<TransactionModelDTO, 'account_from_id' | 'account_to_id' | 'amount' | 'description'>;
