+++ backend/src/repositories/dtos/index.ts
@@ -6,11 +6,3 @@ export * from './Rate/RateInput';
 export * from './Rate/RateModel';
 export * from './Transaction/TransactionInput';
 export * from './Transaction/TransactionModel';
-
-export const operators = {
-  '===': 'equals',
-  '!==': 'not',
-  '>=': 'gte',
-  '<=': 'lte',
-  in: 'in',
-};
